requirements.txt
# Reddit Scraping
praw>=7.7.1
pmaw>=3.0.0  # Pushshift wrapper for historical data

# Data Processing
pandas>=2.0.0
numpy>=2.0.0

# NLP & Sentiment Analysis
nltk>=3.8.0
spacy>=3.7.0
textblob>=0.17.0
vaderSentiment>=3.3.2
transformers>=4.30.0
torch>=2.0.0  # for transformers

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.0

# Utilities
python-dotenv>=1.0.0 # load environment variables
tqdm>=4.65.0  # progress bars
requests>=2.31.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0